Record self-study completion after restoring local verification
Keep official leaderboard work outside the local completion criteria because the private course data and official submission environment are unavailable here. Fix local compatibility so the coding environment can verify the self-study implementation end to end.

Constraint: Official CS336 leaderboard resources require private course data and prescribed cluster/submission environments.

Rejected: Marking the project incomplete solely for inaccessible leaderboard tracks | local self-study deliverables are now verified.

Confidence: high

Scope-risk: narrow

Directive: Do not count private leaderboard tracks as local incompletion unless the required course resources are available.

Tested: coding env A1 47 passed/1 xfailed; A2 16 passed; A3 compileall; A4 21 passed; A5 31 passed; git diff --check

Not-tested: Official Stanford MATH/Paloma leaderboard submissions due to unavailable private course resources

+++ assignment5-alignment/tests/adapters.py
@@ -149,6 +149,7 @@ def run_compute_entropy(logits: torch.Tensor) -> torch.Tensor:
     Returns:
         torch.Tensor shape (batch_size, sequence_length)，每个位置的信息熵。
     """
+    logits = logits.float()
     log_probs = torch.nn.functional.log_softmax(logits, dim=-1)
     probs = torch.exp(log_probs)
     return -(probs * log_probs).sum(dim=-1)
@@ -176,7 +177,7 @@ def run_get_response_log_probs(
             "log_probs": shape (batch_size, sequence_length)，各位置的条件 log p(labels|prefix)。
             "token_entropy" (optional): shape (batch_size, sequence_length)，per-token 熵。
     """
-    logits = model(input_ids).logits
+    logits = model(input_ids).logits.float()
     log_probs_all = torch.nn.functional.log_softmax(logits, dim=-1)
     batch_size, sequence_length, _ = log_probs_all.shape
